requirements.txt
absl-py==1.4.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
filelock==3.9.0
fsspec==2023.5.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
lit==16.0.5
Markdown==3.4.3
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
packaging==23.1
Pillow==9.3.0
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
timm==0.9.2
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.3.4
zipp==3.15.0
